Move AAT tests to a separate source set
