+++ src/aat/java/uk/gov/hmcts/cmc/claimstore/aat/Bootstrap.java
@@ -1,22 +1,20 @@
-package uk.gov.hmcts.cmc.claimstore;
+package uk.gov.hmcts.cmc.claimstore.aat;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
 import io.restassured.RestAssured;
+import io.restassured.config.ObjectMapperConfig;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Primary;
-import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Component;
-import uk.gov.hmcts.cmc.claimstore.aat.TestInstance;
 import uk.gov.hmcts.cmc.claimstore.aat.idam.IdamTestService;
 import uk.gov.hmcts.cmc.claimstore.aat.idam.TestUser;
 import uk.gov.hmcts.cmc.claimstore.services.JwtHelper;
 
 import javax.annotation.PostConstruct;
 
-@Primary
 @Component
-@Profile("aat-tests")
-public class AutomatedAcceptanceTestsSetup implements TestsSetup {
+public class Bootstrap {
 
+    private final ObjectMapper objectMapper;
     private final IdamTestService idamTestService;
     private final JwtHelper jwtHelper;
     private final TestUser testUser;
@@ -26,12 +24,14 @@ public class AutomatedAcceptanceTestsSetup implements TestsSetup {
     private String userId;
 
     @Autowired
-    public AutomatedAcceptanceTestsSetup(
+    public Bootstrap(
+        ObjectMapper objectMapper,
         IdamTestService idamTestService,
         JwtHelper jwtHelper,
         TestUser testUser,
         TestInstance testInstance
     ) {
+        this.objectMapper = objectMapper;
         this.idamTestService = idamTestService;
         this.testUser = testUser;
         this.jwtHelper = jwtHelper;
@@ -41,16 +41,18 @@ public class AutomatedAcceptanceTestsSetup implements TestsSetup {
     @PostConstruct
     public void initialize() {
         RestAssured.baseURI = testInstance.getUri();
+        RestAssured.config = RestAssured.config()
+            .objectMapperConfig(
+                ObjectMapperConfig.objectMapperConfig().jackson2ObjectMapperFactory((cls, charset) -> objectMapper)
+            );
         authenticationToken = "Bearer " + idamTestService.logIn(testUser.getUsername(), testUser.getPassword());
         userId = jwtHelper.getUserId(authenticationToken);
     }
 
-    @Override
     public String getUserAuthenticationToken() {
         return authenticationToken;
     }
 
-    @Override
     public String getUserId() {
         return userId;
     }
