+++ src/aat/java/uk/gov/hmcts/cmc/claimstore/aat/idam/TestUser.java
@@ -2,11 +2,9 @@ package uk.gov.hmcts.cmc.claimstore.aat.idam;
 
 import org.hibernate.validator.constraints.NotBlank;
 import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Component;
 
 @Component
-@Profile("aat-tests")
 @ConfigurationProperties(prefix = "aat.test-user")
 public class TestUser {
