+++ src/aat/resources/environment.properties
@@ -1,0 +1,15 @@
+pdf_service.url = http://dummy-url
+document_management.api_gateway.url = http://dummy-url
+core_case_data.api.url = http://dummy-url
+
+staff-notifications.sender = sender@example.com
+staff-notifications.recipient = recipient@example.com
+
+notifications.frontendBaseUrl = https://civil-money-claims.co.uk
+notifications.govNotifyApiKey = some-gov-uk-notify-api-key
+notifications.templates.email.claimantClaimIssued = claimant-claim-issued-template
+notifications.templates.email.defendantClaimIssued = defendant-claim-issued-template
+notifications.templates.email.defendantResponseIssued = defendant-response=issued-template
+notifications.templates.email.claimantResponseIssued = claimant-response-issued-template
+notifications.templates.email.staffMoreTimeRequested = staff-more-time-requested-template
+notifications.templates.email.defendantMoreTimeRequested = defendant-more-time-requested-template
