--- src/integrationTest/java/uk/gov/hmcts/cmc/claimstore/TestsSetup.java
@@ -1,9 +1,0 @@
-package uk.gov.hmcts.cmc.claimstore;
-
-public interface TestsSetup {
-
-    String getUserAuthenticationToken();
-
-    String getUserId();
-
-}
